+++ Tests/Tests/NSArray_TDTFunctionalAdditions_Tests.m
@@ -33,6 +33,29 @@
   [super tearDown];
 }
 
+- (void)testReductionOfEmptyArrayReturnsInitialObject {
+  id initialObject = @1;
+  id result = [@[] tdt_objectAfterReducingWithInitialObject:initialObject block:^id(id accumulator, id object) {
+    return @2;
+  }];
+  XCTAssertEqualObjects(initialObject, result);
+}
+
+- (void)testInitialObjectIsUsedInFirstCall {
+  id initialObject = @1;
+  [@[@2] tdt_objectAfterReducingWithInitialObject:initialObject block:^id(id accumulator, id object) {
+    XCTAssertEqualObjects(initialObject, accumulator);
+    return nil;
+  }];
+}
+
+- (void)testReduction {
+  id result = [@[@2, @3] tdt_objectAfterReducingWithInitialObject:@1 block:^id(id accumulator, id object) {
+    return @([accumulator intValue] + [object intValue]);
+  }];
+  XCTAssertEqualObjects(@6, result);
+}
+
 - (void)testMapOfEmptyArray {
   XCTAssertEqualObjects([self.emptyArray tdt_arrayByMappingWithSelector:self.identitySelector],
                         self.emptyArray);