+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * Fix variable name in macro.
   * Add method to check blank string.
+  * Add method for array reduction.
 
 2.5.1, 22-July-2014